Info panels: thing and sector flags are now displayed. Visual mode: changed "Move Texture Left/Right/Up/Down by 8" actions to "Move Texture Left/Right/Up/Down by cur. grid size" Visual mode: holding Ctrl-Shift while dragging texture with the mouse will snap texture offsets to current grid size. Classic modes: interiors of highlighted sectors are now highlighted when the action target of currently highlighted map element is sector. Unhandeled exceptions infos are now written to the event log.

+++ branches/GZDoomBuilder/Source/Core/Windows/ThingEditForm.cs
@@ -643,7 +643,7 @@ namespace CodeImp.DoomBuilder.Windows
 			} else { //update values
 				// Apply position
 				foreach(Thing t in things) {
-					float z = (float)posZ.GetResult((int)t.Position.z);
+					float z = posZ.GetResult((int)t.Position.z);
 					if(useAbsoluteHeight && t.Sector != null)
 						z -= t.Sector.FloorHeight;
 					t.Move(new Vector3D(t.Position.x, t.Position.y, z));
